evaluation/requirements.txt
opencv-python
openai
python-dotenv 